refactor(dynamite): unify allOf and object generation
